Move class SimulationPlotter to its own file
